Extracted test mods to separate files, linked with #[path] macro
